implementation

+++ apbd-5-s22085/DTOs/NewProductWarehouseDto.cs
@@ -4,7 +4,7 @@ public record NewProductWarehouseDto(
     int warehouseId,
     int productId,
     int orderId,
-    int amount,
-    int price,
+    int amount, 
+    decimal price,
     DateTime createdAt
 );
